Fix issue where cache kills the world
Not sure what the OSError is caching, but `open` throws an IOError when
it can't open a file.  Not sure what OSError is trying to catch here,
but if it's the file error (which seems like the issue it would be
catching) I don't know how it could have worked.